added daily satisfaction to diary
