+++ src/views/Diary.vue
@@ -2,17 +2,17 @@
 import { onMounted, onBeforeMount } from 'vue'
 import SpinnerLoadingPage from '../components/SpinnerLoadingPage.vue';
 import NoData from '../components/NoData.vue';
-import { spinnerLoadingPage, diaries, selectedItem, spinnerLoadMore, endOfList, tags } from '../stores/globals';
+import { spinnerLoadingPage, diaries, selectedItem, spinnerLoadMore, endOfList, tags, satisfactionArray } from '../stores/globals';
 import { useCheckVisibleScreen, useCreatedDateFormat, useEditItem, useInitPopover, useLoadMore } from '../utils/utils';
 import { useGetDiaries } from '../utils/diary';
-import { useGetTags, useGetTagInfo, useGetAvailableTags } from '../utils/daily';
+import { useGetTags, useGetTagInfo, useGetAvailableTags, useDailySatisfactionChange, useGetSatisfactions } from '../utils/daily';
 
 onBeforeMount(async () => {
 
 })
 
 onMounted(async () => {
-    await Promise.all([useGetDiaries(true), useGetTags(), useGetAvailableTags()])
+    await Promise.all([useGetDiaries(true), useGetTags(), useGetAvailableTags(), useGetSatisfactions()])
     useInitPopover()
     window.addEventListener('scroll', () => {
         let scrollTop = window.scrollY
@@ -42,7 +42,12 @@ onMounted(async () => {
                     <div class="dailyCard quill">
                         <div class="row">
                             <div class="col-12 cardFirstLine d-flex align-items-center fw-bold">
-                                <div class="col-auto">{{ useCreatedDateFormat(itemDiary.dateUnix) }}</div>
+                                <div class="col-auto">{{ useCreatedDateFormat(itemDiary.dateUnix) }}
+                                    <i v-on:click="useDailySatisfactionChange(itemDiary.dateUnix, true)"
+                                        v-bind:class="[(satisfactionArray.find(obj => obj.dateUnix == itemDiary.dateUnix)!=undefined && satisfactionArray.find(obj => obj.dateUnix == itemDiary.dateUnix).satisfaction == true) ? 'greenTrade' : '', 'uil', 'uil-thumbs-up', 'ms-2', 'me-1', 'pointerClass']"></i>
+                                    <i v-on:click="useDailySatisfactionChange(itemDiary.dateUnix, false)"
+                                        v-bind:class="[(satisfactionArray.find(obj => obj.dateUnix == itemDiary.dateUnix) != undefined && satisfactionArray.find(obj => obj.dateUnix == itemDiary.dateUnix).satisfaction == false) ? 'redTrade' : '', , 'uil', 'uil-thumbs-down', 'pointerClass']"></i>
+                                </div>
                                 <span class="col mb-2 ms-auto text-end">
                                     <i class="uil uil-edit-alt pointerClass"
                                         v-on:click="useEditItem(itemDiary.objectId)"></i>
@@ -56,11 +61,12 @@ onMounted(async () => {
                             <div>
                                 <span v-for="tags in tags.filter(obj => obj.tradeId == itemDiary.dateUnix.toString())">
                                     <span v-for="tag in tags.tags.slice(0, 7)" class="tag txt-small"
-                                        :style="{ 'background-color': useGetTagInfo(tag).groupColor }">{{useGetTagInfo(tag).tagName }}
+                                        :style="{ 'background-color': useGetTagInfo(tag).groupColor }">{{ useGetTagInfo(tag).tagName
+                                        }}
                                     </span>
                                     <span v-show="tags.tags.length > 7">+{{
                                         tags.tags.length
-                                        -  7}}</span>
+                                        - 7 }}</span>
                                 </span>
                             </div>
                             <div class="col-12 mt-2">
